requirements.txt
flask
flask_sqlalchemy
flask_cors
flask_migrate
flask_jwt_extended 
psycopg2
boto3
python-dotenv
gunicorn
werkzeug